Updated the engine. Fixed image problem creating 4 different images instead of 1 and then drawing regions. Fixed player positioning issue.

+++ src/main/java/com/gjkf/lunarLander/core/gui/screens/MainScreen.java
@@ -15,7 +15,7 @@ public class MainScreen extends GuiScreenWidget{
     private Player player;
     private Terrain terrain;
 
-    public MainScreen(int width, int height){
+    public MainScreen(float width, float height){
         super(width, height);
         this.startTime = System.currentTimeMillis();
         player = new Player(100, 100, 32, 32);
